fixed modal sizing issue for mp and qa
